perf(user): 重构user模块路径，在注册中添加了角色鉴权
