+++ .env
@@ -1,0 +1,1 @@
+DATABASE_URL=postgres://username:password@localhost/diesel_demo
